Medium Support (Multi-disc Albums), Quality Grouping (#121)
* Multi Disc Stage 1 - Backend Work

* Quality Group Functionality

* Fixed: Only show wanted album types on ArtistDetail page

* Add Media Count Column to ArtistDetail Page

* Parser updates for multidisc cases, other usenet release title formats

* Search for Tracks by Medium Number in Addition to Title and TrackNumber

* Medium Renaming Token for Track Naming

* fixup Codacy and Comment Cleanup

* fixup remove comments

+++ src/NzbDrone.Core/Profiles/Quality/ProfileService.cs
@@ -73,7 +73,9 @@ namespace NzbDrone.Core.Profiles.Qualities
                             .Select(v => new ProfileQualityItem { Quality = v.Quality, Allowed = allowed.Contains(v.Quality) })
                             .ToList();
 
-            var profile = new Profile { Name = name, Cutoff = cutoff, Items = items};
+            var profile = new Profile { Name = name,
+                Cutoff = (int)cutoff,
+                Items = items};
 
             return Add(profile);
         }
